[MCUX-73496] Resume the pended L2CAP channel connection REQ
There are two cases if the L2CAP channel connect REQ is received,
Case 1, the ACL connection has been encrypted, the status `SUCCESS`
is returned.
Case 2, the ACL connection is not encrypted, the status `connection
pending` is returned.

For case 1, there is not any issues.
For case 2, there are two situations,
Situation 1, if the BT Settings is not enabled, the ACL connection
is encrypted before connection REQ is handled by Edgefast Host.
Situation 2, if the BT Settings is enabled, the ACL connection is
not encrypted before connection REQ is handled by Edgefast Host. The
L2CAP connection is in pending status. And after the encryption
changed event triggered, the connection procedure should be resumed.

The issue occurs in case 2 situation 2. The pended L2CAP channel
connection REQ is not resumed.

Fix the issue by resuming the pended L2CAP channels.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>