[MCUX-73548] CONFIG_WORK_QUEUE_TASK_PRIORITY=(configMAX_PRIORITIES-2)
There is an issue that the EdgeFast Bluetooth Host task is blocked
by EtherMind thask. It causes the timing of EdgeFast Bluetooth Host
is broken.

Now, the priority of  EdgeFast Bluetooth Host task is `4`. And the
priority of EtherMind task is `configMAX_PRIORITIES - 3`.

Fix the issue by changing CONFIG_WORK_QUEUE_TASK_PRIORITY to
`configMAX_PRIORITIES - 2`.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>