[MCUX-73555] Optimize resource release
There is a additional issue when the ACL connection is broken before
the ISO connection disconnected that the resources of BAP/ISO are
not free/cleared properly.

In the current implementation, the stream/cis/cig cannot be attached
before all streamer are released. Otherwise, the logic of BAP will
be broken, and the potential issue is that any some resources will
not be cleared/free.

Optimize the ACL disconnection event with following steps,
Step 1, flag the ACL disconnecting state,
Step 2, check if the streamer has been released. If it has been
released, free the resources of application for this streamer,
Step 3, check if all streamers of current ACL connection are
released. If it is positive, free the resources of application for
this ACL connection,
Step 4, check if all ACL connections are released. If it is positive,
try to delete the group.

For the case, the streamer is not released when ACL connection is
broken. The following steps is added in streamer released event
callback,
Step 1, check if the ACL of the streamer is broken. If it is positive,
free the resources of application for this streamer,
Step 2, check if all streamers of any ACL connection are released. If
it is positive, free the resources of application for this ACL
connection,
Step 3, check if all ACL connections are released. If it is positive,
try to delete the group.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>