[MCUX-73584] Deinit Codec in Codec task context to avoid transfer issue
There is an issue that the Codec task is working on sending audio
data via I2S. Meanwhile, there is a high priority thread preempts
the CPU to de-initialize the Codec. Than after the codec tack gets
the CPU, the issue occurs in I2S data sending, because the I2S has
been de-initialized.

Add a flag for de-initializing the Codec. And then in Codec task,
if the flag is true, and then call Codec de-initialization function.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>