[MCUX-73961][RT700][Edgefast_bluetooth][MCUX]hardfault in LFS init
LFS accesses memory through Mflash and XSPI. XSPI driver will reset AHB
bus when XSPI_ReadBlocking() is timeout. Currently, text of Mflash and
XSPI is placed in external flash which is mounted on AHB bus. When AHB
is reseted by XSPI driver, HARDFAULT occurred.

With this patch, both text and rodata of Mflash and XSPI are placed in
RAM.

Signed-off-by: Jiawei Yang <jiawei.yang_1@nxp.com>