[MCUX-73585][SDK_24.12.00-pvw2_RC1][LE_Audio][unicast_media_sender/UMS_a2dp_bridge] failed to set security --1/10
When bt_conn_set_security() fails on the central side, it doesn't mean that security will eventually fail.
It could just be that there is already a process setting security. If we don't wait for the security event, all the next processes will fail.
So we still need to wait for the security change event.

With the patch, central will always wait security change event.