[MCUX-74073] Increase the priority of timer task
In current settings, the consistency of FreeRTOS timer object can not
be ensured, since the priorities of some tasks are higher than the
priority of timer task.

Increase the priority of the timer task from "configMAX_PRIORITIES - 3"
to "configMAX_PRIORITIES - 1".

It can ensure the timer task with the highest priority.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>