[MCUX-74083] Add a flag to flag the timer status
There is an issue that the time expiry callback will be called
multiple times if the timer has been expired with multiple times
timeout.
Then the timer will be stopped multiple times. Than the timer object
queue will be broken.
Then the FreeRTOS will be hang in timer stop sending queue.

And a flag K_WORK_DELAY_Q_TIMER to flag the timer status.

Only when the timer is enabled, then stop the timer.
Only when the timer is disabled, then start the timer.

Signed-off-by: Lyle Zhu <lyle.zhu@nxp.com>