[MCUX-74847][LE Audio][RT700] update iar linker file put lc3 into ram.
There is a lib name changed from libethermind_ble_lc3_xxx to
libethermind_lc3_xxx, but linker not updated. So that the lc3 not in
ram on rt700 le audio iar target. And this will case performance down.

In this patch, iar linker file updated to put lc3 into ram.

Signed-off-by: Chang An <chang.an_1@nxp.com>